requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
langchain-cohere
python-dotenv
cohere